+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.2.5
+* Dropped endpoint.Addr entity as unused. After change link type between session and endpoint 
+  to NodeID endpoint.Addr became unnecessary for internal logic of driver
+
 ## 3.2.4
 * Returned context error when context is done on `session.StreamExecuteScanQuery` 
   and `session.StreamReadTable`